requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
SQLAlchemy==2.0.23
psycopg[binary]==3.1.18
python-dotenv==1.0.0
Jinja2==3.1.2
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.27.2
openai==1.40.0
python-multipart==0.0.9
alembic==1.13.0

